requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
PyMuPDF
numpy
streamlit
openai
